+++ cloudos_cli/analytics/__init__.py
@@ -1,0 +1,8 @@
+"""
+Functions and classes related to analytics.
+"""
+
+from .analytics import Analytics
+
+
+__all__ = ['Analytics']